Backend/requirements.txt
torch>=1.7.1
torchvision>=0.8.2
open_clip_torch>=2.0.2
transformers>=4.18.0
tqdm>=4.64.0
numpy>=1.19.5
pillow>=8.0.0
matplotlib>=3.3.4
tensorboard>=2.14.0
wordcloud>=1.8.0
nltk>=3.6.0
textblob>=0.15.0
accelerate>=0.20.0
timm>=0.6.0
fairscale>=0.4.0
sentencepiece>=0.1.97
datasets>=2.10.0
pandas>=1.3.0
rouge-score>=0.1.2
av>=10.0.0
huggingface_hub>=0.20.0 